MARS-1169 Split test suite
