+++ client/test/integration/actions/dashboard.test.ts
@@ -4,7 +4,7 @@ import test, { expect } from "@playwright/test";
 // Test helper functions
 import { createTestUser, createTestWorkspace, switchWorkspace } from "../helpers";
 
-test.describe("Interface launches", () => {
+test.describe("Dashboard", () => {
   test.beforeEach(async ({ context, page }) => {
     // Create User
     const user = await createTestUser(context);
@@ -17,7 +17,7 @@ test.describe("Interface launches", () => {
     await switchWorkspace(page, "Dashboard-1");
   });
 
-  test("navigation menu items are visible", async ({ page }) => {
+  test("navigation items are visible", async ({ page }) => {
     await page.goto("/");
     await expect(page.locator("#navSearchButtonDesktop")).toHaveText("Search");
     await expect(page.locator("#navProjectsButtonDesktop")).toHaveText("Projects");
