+++ client/test/integration/actions/project.test.ts
@@ -7,7 +7,6 @@ import {
   openItemFromTable,
   saveAndWait,
   clickButtonByText,
-  getUniqueName,
   createTestEntity,
   createTestProject,
   createTestUser,
@@ -17,55 +16,7 @@ import {
 } from "../helpers";
 
 test.describe("Project", () => {
-  test.describe("Create", () => {
-    test.beforeEach(async ({ context, page }) => {
-      // Create User
-      const user = await createTestUser(context);
-
-      // Setup Workspace
-      await createTestWorkspace("Project-1", user);
-
-      // Setup navigation
-      await switchWorkspace(page, "Project-1");
-      await page.goto("/create/project");
-      await expect(page.locator("h2:has-text('Create Project')")).toBeVisible();
-    });
-
-    test("should create a project with required fields", async ({ page }) => {
-      await page.locator("[data-testid='create-project-name']").fill("1-Project-Create");
-      await page.locator('input[type="datetime-local"]').fill("2023-10-01T12:00");
-      await page.locator("[data-testid='create-project-description']").fill("This is a test project description.");
-      await page.click("[data-testid='create-project-finish']");
-
-      await expect(page).toHaveURL(/\/projects/);
-    });
-
-    test("should validate required fields", async ({ page }) => {
-      await expect(page.locator("[data-testid='create-project-finish']")).toBeDisabled();
-
-      await page.locator("[data-testid='create-project-name']").fill(getUniqueName("Project Name"));
-      await expect(page.locator("[data-testid='create-project-finish']")).toBeDisabled();
-
-      await page.locator("[data-testid='create-project-description']").fill("Test description");
-      await page.locator('input[type="datetime-local"]').fill("2023-10-01T12:00");
-      await expect(page.locator("[data-testid='create-project-finish']")).toBeEnabled();
-    });
-
-    test("should handle cancel button", async ({ page }) => {
-      await page.locator("[data-testid='create-project-name']").fill(getUniqueName("Project Cancel"));
-      await page.click("[data-testid='create-project-cancel']");
-
-      // A blocker dialog may appear when navigating away with unsaved changes
-      const continueButton = page.locator('button:has-text("Continue")');
-      if ((await continueButton.count()) > 0) {
-        await continueButton.click();
-      }
-
-      await expect(page).toHaveURL(/\/projects/);
-    });
-  });
-
-  test.describe("Edit details", () => {
+  test.describe("Edit: Details", () => {
     test.beforeEach(async ({ context, page }) => {
       // Create User
       const user = await createTestUser(context);
@@ -110,7 +61,7 @@ test.describe("Project", () => {
     });
   });
 
-  test.describe("Edit Entities", () => {
+  test.describe("Edit: Entities", () => {
     test.beforeEach(async ({ context, page }) => {
       // Create User
       const user = await createTestUser(context);
@@ -170,7 +121,7 @@ test.describe("Project", () => {
     });
   });
 
-  test.describe("List visibility", () => {
+  test.describe("View", () => {
     test.beforeEach(async ({ context, page }) => {
       // Create User
       const user = await createTestUser(context);
