+++ client/test/integration/security/endpoints.test.ts
