+++ client/test/integration/workflows/import.test.ts
@@ -33,7 +33,7 @@ test.describe("Import", () => {
       await clickButtonWhenEnabled(page, "#navImportButtonDesktop");
 
       // Upload CSV file
-      const csvPath = path.resolve(process.cwd(), "test/playwright/fixtures/export_entities.csv");
+      const csvPath = path.resolve(process.cwd(), "test/integration/export_entities.csv");
       const fileInput = page.locator('input[type="file"]').first();
       await fileInput.setInputFiles(csvPath);
 
@@ -63,7 +63,7 @@ test.describe("Import", () => {
       await clickButtonWhenEnabled(page, "#navImportButtonDesktop");
 
       // Upload JSON file
-      const jsonPath = path.resolve(process.cwd(), "test/playwright/fixtures/export_entities.json");
+      const jsonPath = path.resolve(process.cwd(), "test/integration/export_entities.json");
       const fileInput = page.locator('input[type="file"]').first();
       await fileInput.setInputFiles(jsonPath);
